backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langgraph==0.1.9
asyncpg==0.29.0
redis==5.0.4
python-multipart==0.0.9
pypdf==4.2.0
python-dotenv==1.0.1
pydantic==2.7.4
google-genai==1.0.0
